[HUDI-5434] Fix archival in metadata table to not rely on completed rollback or clean in data table (#7580)
Before this change, the archival for the metadata table uses the earliest instant of all actions from the active timeline of the data table. In the archival process, CLEAN and ROLLBACK instants are archived separately apart from commits (check HoodieTimelineArchiver#getCleanInstantsToArchive). Because of this, a very old completed CLEAN or ROLLBACK instant in the data table can block the archive of the metadata table timeline and causes the active timeline of the metadata table to be extremely long, leading to performance issues for loading the timeline.

This commit changes the archival in metadata table to not rely on completed rollback or clean in data table, by archiving the metadata table's instants after the earliest commit (COMMIT, DELTA_COMMIT, and REPLACE_COMMIT only, considering non-savepoint commit only if enabling archive beyond savepoint) and the earliest inflight instant (all actions) in the data table's active timeline.

(cherry picked from commit 7a9aabd9b2595158a2610be12d567df4d974e258)

# Conflicts:
#	hudi-client/hudi-client-common/src/main/java/org/apache/hudi/client/HoodieTimelineArchiver.java
#	hudi-common/src/main/java/org/apache/hudi/common/table/timeline/TimelineUtils.java
#	hudi-common/src/test/java/org/apache/hudi/common/table/TestTimelineUtils.java